[FIX] web_editor: handle we-list trailing space
Reproduce issue:
- Drop and drop a form snippet
- Set a field as a multiple checkbox
- Toggle on option 1
- Set a space at the end of option 1 so "option 1 "
- Blur the input

Result:
> You will find, Option 1 is no more selected.

The root cause of this issue appears to be the unexpected trimming of
the HTML input box. By default, HTML trims leading and trailing spaces,
as well as considers multiple inner spaces as a single space.
Consequently, inconsistencies arise between the input value, ID, and
display name. This commit trims the value from the initial to maintain
consistency.

As a result, `option 1`, `option 1   `, `  option 1`, and `option     1`
will all be trimmed and considered the same.

task-3254884

closes odoo/odoo#170520

X-original-commit: 5e67f4a7911402a7edec4a477157baf1f44de570
Signed-off-by: Robin Lejeune (role) <role@odoo.com>