[FIX] account: allow credit card accounts on bank journals
Currently people wanting to use a bank journal for their credit card
transactions have to use a hack to set the default account as a credit
card account, since the domain does not allow this.

To make the life of people easier, we now allow selecting a credit card
account on bank journals.

task-3891250

closes odoo/odoo#167188

X-original-commit: cea1e3865cdeeca5b3979cdf814a53d8d170806d
Related: odoo/enterprise#63437
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>